Blind SSRF via feed parser accessing an internal URL
Vulnerability
By adding an RSS feed portlet in their dashboard, a normal member could try loading the RSS feed of an internal service that is otherwise unreachable for this member.

Current status as on 20 May, 2020
Hotfixed. Note: the only fix we do, is that we only allow HTTP and HTTPS access, so you cannot use the file protocol or any other protocol. The more general case of accessing internal servers here and in other parts of the Plone code, cannot truly be fixed without major code changes or breaking valid use cases. For example, on an Intranet, you may want to show the RSS feed of an internal log server.